refactor: 책임별로 파일 분리 진행
