+++ app/components/Authentication/AuthenticationCardHeader.tsx
@@ -1,0 +1,15 @@
+import { CardHeader, CardTitle } from "@/components/ui/card";
+
+interface AuthenticationCardHeaderProps {
+  text: string;
+}
+
+export default function AuthenticationCardHeader({
+  text,
+}: AuthenticationCardHeaderProps) {
+  return (
+    <CardHeader>
+      <CardTitle className="text-center">{text}</CardTitle>
+    </CardHeader>
+  );
+}
